ml_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
numpy==1.24.3
pandas==2.0.3
catboost==1.2.2
torch==2.1.0

